Refactor handler codes
