+++ test/index.test.js
@@ -1,10 +1,11 @@
 jest.mock('@octokit/rest')
 
 const Octokit = require('@octokit/rest')
-const index = require('../src/index')
+const { handler } = require('../src/index')
+const { createIssue } = require('../src/handler')
 
 describe('handler', () => {
-  test('responds with the status code 200', async () => {
+  test('responds with the status code 200 if completes the process', async () => {
     process.env.GITHUB_TOKEN = 'githubToken'
 
     Octokit.mockImplementation(() => ({
@@ -23,19 +24,41 @@ describe('handler', () => {
     }))
 
     const event = {
-      "body": 'user_name=kentaro&text=test+issue'
+      'body': 'user_name=kentaro&text=test+issue'
     }
 
-    const response = await index.handler(event)
+    const response = await handler(event)
 
     expect(response.statusCode).toEqual(200)
 
     process.env.GITHUB_TOKEN = ''
   })
 
-  test('creates a message to notify Slack', async () => {
-    process.env.GITHUB_TOKEN = 'githubToken'
+  test('responds with the status code 500 if throws an error', async () => {
+    process.env.GITHUB_TOKEN = ''
 
+    Octokit.mockImplementation(() => ({
+      authenticate: jest.fn().mockImplementation(async () => {
+        throw new Error('Bad credentials')
+      })
+    }))
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await handler(event)
+
+    expect(response.statusCode).toEqual(500)
+  })
+})
+
+describe('createIssue', () => {
+  beforeEach(() => {
+    process.env.GITHUB_TOKEN = 'githubToken'
+  })
+
+  test('creates a message to notify Slack', async () => {
     Octokit.mockImplementation(() => ({
       authenticate: jest.fn(),
       issues: {
@@ -51,11 +74,33 @@ describe('handler', () => {
       }
     }))
 
-    const event = {
-      "body": 'user_name=kentaro&text=test+issue'
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
     }
 
-    const response = await index.handler(event)
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
 
     expect(response.statusCode).toEqual(200)
     expect(JSON.parse(response.body)).toEqual({
@@ -78,25 +123,242 @@ describe('handler', () => {
         }
       ]
     })
-
-    process.env.GITHUB_TOKEN = ''
   })
 
-  test('responds with the status code 500 if GITHUB_TOKEN is not set', async () => {
-    process.env.GITHUB_TOKEN = ''
-
+  test('sets a base url if uses on the GitHub Enterprise', async () => {
     Octokit.mockImplementation(() => ({
-      authenticate: jest.fn().mockImplementation(async () => {
-        throw new Error('Bad credentials')
-      })
+      authenticate: jest.fn(),
+      issues: {
+        create: jest.fn().mockImplementation(async () => {
+          return {
+            data: {
+              number: 1234,
+              title: 'Found a bug',
+              html_url: 'https://github.com/kentaro-m/issue-creator/issues/1234'
+            }
+          }
+        })
+      }
     }))
 
-    const event = {
-      "body": 'user_name=kentaro&text=test+issue'
+    const AppConfig = {
+      'github': {
+        'baseUrl': 'https://githubenterprise.com/api/v3',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
     }
 
-    const response = await index.handler(event)
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(200)
+    expect(Octokit.mock.calls[2][0].baseUrl).toEqual('https://githubenterprise.com/api/v3')
+
+  })
+
+  test('throws an error if repoOwner is not set', async () => {
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': '',
+        'repoName': 'idea'
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
 
     expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: No repoOwner found on the configuration file.')
+  })
+
+  test('throws an error if repoName is not set', async () => {
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': ''
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: No repoName found on the configuration file.')
+  })
+
+  test('throws an error if GITHUB_TOKEN is not set', async () => {
+    process.env.GITHUB_TOKEN = ''
+
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: No GITHUB_TOKEN found on environment variables.')
+  })
+
+  test('throws an error if a users key is not exist', async () => {
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      }
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: users is not iterable')
+  })
+
+  test('throws an error if a users list is not set', async () => {
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      },
+      'users': []
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text=test+issue'
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: Not permitted to execute the issue creator command.')
+  })
+
+  test('throws an error if an issue title is not set', async () => {
+    Octokit.mockImplementation(() => ({
+      authenticate: jest.fn(),
+      issues: {
+        create: jest.fn().mockImplementation(async () => {
+          return {
+            data: {
+              number: 1234,
+              title: 'Found a bug',
+              html_url: 'https://github.com/kentaro-m/issue-creator/issues/1234'
+            }
+          }
+        })
+      }
+    }))
+
+    const AppConfig = {
+      'github': {
+        'baseUrl': '',
+        'repoOwner': 'kentaro-m',
+        'repoName': 'idea'
+      },
+      'users': [
+        {
+          'slack': 'john',
+          'github': 'john'
+        },
+        {
+          'slack': 'debra',
+          'github': 'debra'
+        },
+        {
+          'slack': 'kentaro',
+          'github': 'kentaro-m'
+        }
+      ]
+    }
+
+    const event = {
+      'body': 'user_name=kentaro&text='
+    }
+
+    const response = await createIssue(event, AppConfig)
+
+    expect(response.statusCode).toEqual(500)
+    expect(JSON.parse(response.body).attachments[0].text).toEqual('Error: Invalid the issue title.')
   })
 })
